Add Cargo.toml files
The Rust code in this repository would be much more useful if it
could be built with Cargo.

This commit adds Cargo.toml files for various packages in this
repository. At the root of the repository is a virtual manifest
listing every crate in the repo. So you can use `cargo build --all`
to build everything. This manifest has all items which could not
be built commented out.

I tried to keep the Cargo.toml files simple. They are basically
limited to name, version, authors, and license (the latter 3 are
constant across all packages) plus dependencies.

For dependencies, I like when my build systems are deterministic
over time, so I pinned packages to explicit versions.

I didn't attempt to define the types of crates.

There is some code in the repo we don't attempt to compile. For
example, the file layout under cmds/ does not agree with Cargo's
standard layout. I'm not sure if that should be changed to conform
or what. I just ignored scenarios like that for now.

I built with rustc 1.20.0 stable with RUSTC_BOOTSTRAP=1 to enable
nightly features. Even with this suggested configuration, I got some
build failures.

There are a handful of failures due to missing crates. I suspect
many of these are Facebook internal crates. I don't think there's
much I can do to address this.

For the crates that don't build, dependencies are likely incomplete.
Rather than audit the source and guess dependencies, I left things
blank.

On my system, 21 of 34 crates build. That's a start!