storage/types: add Cargo.toml file
This series is a part of making mononoke build using cargo which is rust's
package manager. Initial patches of this series will add Cargo.toml files in
different crates and making them build individually with cargo and
then in a later patch we will introduce a top-level Cargo.toml file with
all the crates listed in workspace.

All the builds are done using rustc version 1.27.0-dev which is build from
source to a stage2 compiler.

storage/types crate builds successfully using cargo.